requirements.txt
numpy
opencv-python
pyzmq
aiohttp
aiohttp-cors
aiortc
av
pygame
pyaudio
sentence-transformers
torch
google-genai
fastapi
uvicorn
requests
python-dotenv
RealtimeSTT
dlib
openwakeword